pyproject.toml
[tool.poetry]
name = "fish"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pandas = "^1.1.4"
torch = "^1.7.0"
toolz = "^0.11.1"
tqdm = "^4.54.1"
object-detection = {path = "object-detection"}
torchvision = "^0.8.1"
requests = "^2.25.0"
Pillow = "^8.0.1"
opencv-python = "^4.4.0"
albumentations = "^0.5.2"
Flask = "^1.1.2"
Flask-Cors = "^3.0.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.790"
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
